requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
more-itertools==9.1.0
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
py==1.11.0
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
twilio==8.4.0
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
